Update deploy-an-angular-site.mdx
Hotfix for Update Build Settings For angular. Fixed and working.

Request to change PagesBuildPreset to this solution or add instructions to change the ng build command in package.json.

Request to update the Angular preset on the dashboard to match. 

Seems like there are quite a few people with this issue.

See: https://github.com/cloudflare/workers-sdk/issues/5527
Also: https://stackoverflow.com/questions/78463333/fail-to-deploy-angular-project-on-cloudflare-pages/79151848#79151848 
And: https://community.cloudflare.com/t/angular-app-deployed-to-pages-is-unavailable-on-domain/632745